// src/modules/comic/comic.response.interface.ts
import { Comic } from "@/modules/comic/comic.dto";

export interface ComicResponse {
  code: number;
  status: string;
  copyright: string;
  attributionText: string;
  attributionHTML: string;
  etag: string;
  data: Data;
}

export interface Data {
  offset: number;
  limit: number;
  total: number;
  count: number;
  results: Comic[];
}